+++ CHANGELOG.md
@@ -1,9 +1,22 @@
-# Change Log
+# ChangeLog
 
 All notable changes to the "flutter-friflex" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.2] - 2025-04-27
 
-- Initial release
+### Added
+- Automatic documentation comments for Dart classes using {@template} tag
+- Support for adding comments to constructors with {@macro} tag
+- Added keyboard shortcuts
+
+## [0.0.1] - 2025-04-25
+
+### Added
+- Initial release of Flutter Friflex extension
+- Command to create feature folders structure
+- Command to generate Flutter widgets
+- Command to generate entity classes
+- Command to add documentation comments to Dart classes